LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-spacing-trim/text-spacing-trim-combinations-001_class=vrl&test=MO_FH-expected.txt


FAIL 0: ・《 assert_equals: expected "FH" but got "FF"
FAIL 1: ・「 assert_equals: expected "FH" but got "FF"
FAIL 2: ・『 assert_equals: expected "FH" but got "FF"
FAIL 3: ・〔 assert_equals: expected "FH" but got "FF"
FAIL 4: ・〖 assert_equals: expected "FH" but got "FF"
FAIL 5: ・〘 assert_equals: expected "FH" but got "FF"
FAIL 6: ・〝 assert_equals: expected "FH" but got "FF"
FAIL 7: ・（ assert_equals: expected "FH" but got "FF"
FAIL 8: ・｛ assert_equals: expected "FH" but got "FF"
FAIL 9: ・｟ assert_equals: expected "FH" but got "FF"
FAIL 10: 　《 assert_equals: expected "FH" but got "FF"
FAIL 11: 　「 assert_equals: expected "FH" but got "FF"
FAIL 12: 　『 assert_equals: expected "FH" but got "FF"
FAIL 13: 　〔 assert_equals: expected "FH" but got "FF"
FAIL 14: 　〖 assert_equals: expected "FH" but got "FF"
FAIL 15: 　〘 assert_equals: expected "FH" but got "FF"
FAIL 16: 　〝 assert_equals: expected "FH" but got "FF"
FAIL 17: 　（ assert_equals: expected "FH" but got "FF"
FAIL 18: 　｛ assert_equals: expected "FH" but got "FF"
FAIL 19: 　｟ assert_equals: expected "FH" but got "FF"
FAIL 20: ：《 assert_equals: expected "FH" but got "FF"
FAIL 21: ：「 assert_equals: expected "FH" but got "FF"
FAIL 22: ：『 assert_equals: expected "FH" but got "FF"
FAIL 23: ：〔 assert_equals: expected "FH" but got "FF"
FAIL 24: ：〖 assert_equals: expected "FH" but got "FF"
FAIL 25: ：〘 assert_equals: expected "FH" but got "FF"
FAIL 26: ：〝 assert_equals: expected "FH" but got "FF"
FAIL 27: ：（ assert_equals: expected "FH" but got "FF"
FAIL 28: ：｛ assert_equals: expected "FH" but got "FF"
FAIL 29: ：｟ assert_equals: expected "FH" but got "FF"
・《
・「
・『
・〔
・〖
・〘
・〝
・（
・｛
・｟
　《
　「
　『
　〔
　〖
　〘
　〝
　（
　｛
　｟
：《
：「
：『
：〔
：〖
：〘
：〝
：（
：｛
：｟
